feat: Implement user authentication with Google login and OTP verification
- Added middleware to check user credentials before sending OTP in authRoutes.
- Updated userRoutes to include Google login and modified login endpoint to use OTP.
- Refactored createCookie utility to only require sessionId.
- Introduced createUserWithEssentials utility for user creation with essential data.
- Enhanced recursiveDirectoryRemover to prevent errors when deleting non-existent files.
- Updated sendOtpService to include email in the mailer configuration.
- Modified setupDB to adjust user schema and include new fields for authentication.
- Added verifyIdToken utility for Google ID token verification.
- Integrated Google OAuth in Frontend with new dependencies and components.
- Updated Login and Register components to include Google login button.
- Enhanced DirectoryHeader to display user profile image if available.
- Created ProfileImage component for rendering user profile images.

+++ Backend/utils/recursiveDirectoryRemover.js
@@ -5,15 +5,13 @@ import File from "../models/fileModel.js";
 
 export const deleteDirRecursively = async (currDirId) => {
   // get all the sub-directories of current directory
-  const directories = await Directory.find(
-    { parentDir: currDirId },
-    { projection: { _id: 1 } },
-  );
+  const directories = await Directory.find({ parentDir: currDirId })
+    .select("_id")
+    .lean();
   // get all the files of the current directory
-  const files = await File.find(
-    { parentDir: currDirId },
-    { projection: { extname: 1 } },
-  );
+  const files = await File.find({ parentDir: currDirId })
+    .select("extname")
+    .lean();
 
   // remove all the sub-directories and their contents recursively
   for (const { _id: dirId } of directories) await deleteDirRecursively(dirId);
@@ -22,9 +20,17 @@ export const deleteDirRecursively = async (currDirId) => {
   for (const { _id: fileId, extname } of files) {
     await fs.rm(path.join(process.cwd(), "storage", fileId + extname));
   }
+
   // remove all the files from DB where parent is 'dir'
-  await File.deleteMany({ parentDir: currDirId });
+  if (files.length) await File.deleteMany({ parentDir: currDirId });
 
   // remove the curr-dir from Directory now
   await Directory.findByIdAndDelete(currDirId);
 };
+
+// [Error: ENOENT: no such file or directory, lstat '/mnt/data1/my-files/Projects/Personal-Google-Drive/Backend/storage/6990a8018cf2fad472d40ae8undefined'] {
+//   errno: -2,
+//   code: 'ENOENT',
+//   syscall: 'lstat',
+//   path: '/mnt/data1/my-files/Projects/Personal-Google-Drive/Backend/storage/6990a8018cf2fad472d40ae8undefined'
+// }
